requirements.txt
astroid==2.2.5
cached-property==1.5.1
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
coverage==4.5.3
Django==2.2.2
django-cors-headers==3.0.2
django-downloadview==1.9
django-filter==2.1.0
django-nose==1.4.6
djangorestframework==3.9.4
entrypoints==0.3
enum-compat==0.0.2
flake8==3.7.7
flake8-polyfill==1.0.2
future==0.17.1
geographiclib==1.49
geopy==1.20.0
gunicorn==19.9.0
idna==2.8
isort==4.3.20
joblib==0.13.2
lazy-object-proxy==1.4.1
mando==0.6.4
mccabe==0.6.1
nose==1.3.7
numpy==1.16.4
pandas==0.25.0
polyline==1.4.0
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pylint==2.3.1
PyMySQL==0.9.3
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.1
radon==3.0.3
redis==3.2.1
requests==2.22.0
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
